+++ blocklets/ai-studio/CHANGELOG.md
@@ -1,6 +1,6 @@
 ## 0.1.157 (2023-12-6)
 
-- chore: adjust the project and settings ui
+- fix: adjust the project and settings ui
 
 ## 0.1.156 (2023-12-5)